==> 01_enem_2016.ipynb <==
import pandas as pd
import zipfile

colunas = ['NU_INSCRICAO','NU_ANO', 'CO_MUNICIPIO_RESIDENCIA', 'SG_UF_RESIDENCIA', 'NU_IDADE', 'TP_SEXO', 'TP_ESTADO_CIVIL', 'TP_COR_RACA', 'TP_NACIONALIDADE', 'TP_ST_CONCLUSAO', 'TP_ANO_CONCLUIU', 'TP_ESCOLA', 'TP_ENSINO', 'IN_TREINEIRO',
          'CO_ESCOLA','CO_UF_ESC', 'NO_MUNICIPIO_ESC', 'SG_UF_ESC', 'TP_DEPENDENCIA_ADM_ESC', 'TP_LOCALIZACAO_ESC', 'TP_SIT_FUNC_ESC',
          'TP_PRESENCA_CN', 'TP_PRESENCA_CH', 'TP_PRESENCA_LC', 'TP_PRESENCA_MT', 'NU_NOTA_CH','NU_NOTA_CN','NU_NOTA_LC','NU_NOTA_MT','TP_LINGUA',
          'NU_NOTA_COMP1','NU_NOTA_COMP2','NU_NOTA_COMP3','NU_NOTA_COMP4','NU_NOTA_COMP5','NU_NOTA_REDACAO','TP_STATUS_REDACAO',
          'Q001','Q002','Q003','Q004','Q005','Q006','Q007','Q008','Q009','Q010','Q011','Q012','Q013','Q014','Q015','Q016','Q017','Q018','Q019','Q020','Q021','Q022','Q023','Q024','Q025']

ano = '2016'
path_original ='D:/000000001_ENEM/microdados_enem_'+ano+'.zip'

with zipfile.ZipFile(path_original) as z:
    print(*z.namelist(), sep = '\n')

with zipfile.ZipFile(path_original) as z:
    with z.open('Microdados_enem_2016/DADOS/microdados_enem_2016.csv') as f:
        enem_ano = pd.read_csv(f, sep = ';', encoding = 'ISO-8859-1', usecols = colunas)

path_out = 'D:/000000001_ENEM/DADOS_PROCESSADOS/'

medias_ano = enem_ano[['NU_NOTA_REDACAO','NU_NOTA_CN', 'NU_NOTA_MT', 'NU_NOTA_LC', 'NU_NOTA_CH', 'NU_ANO']].groupby(['NU_ANO']).mean().reset_index()

medias_tpescola = enem_ano[['NU_NOTA_REDACAO','NU_NOTA_CN', 'NU_NOTA_MT', 'NU_NOTA_LC', 'NU_NOTA_CH', 'NU_ANO', 'TP_ESCOLA']].groupby(['NU_ANO', 'TP_ESCOLA']).mean().reset_index()

medias_ano.to_csv(path_out+'medias_'+ano+'.csv', sep = ';', encoding = 'ISO-8859-1', index = None)

medias_tpescola.to_csv(path_out+'medias_tpescola_'+ano+'.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

enem_ano['SG_UF_ESC'].fillna('XX',inplace = True)

lista_ufs = ['RS','DF','MT']

for uf in lista_ufs:
    path = path_out+'enem_'+ano+'_{}.csv'.format(uf)
    enem_ano[enem_ano['SG_UF_ESC'] == uf].to_csv(path, sep = ';', index = None, encoding = 'ISO-8859-1')
