==> 05_concatenar.ipynb <==
import pandas as pd

path = 'D:/000000001_ENEM/DADOS_PROCESSADOS/'

anos = ['2015', '2016', '2017', '2018', '2019']

concatenado_rs = pd.DataFrame()
concatenado_mt = pd.DataFrame()
concatenado_df = pd.DataFrame()
medias_gerais = pd.DataFrame()
medias_gerais_tpescola = pd.DataFrame()
concatenado = pd.DataFrame()

for a in anos:
    enem_rs = pd.read_csv(path+'enem_'+a+'_RS.csv', sep = ';', encoding = 'ISO-8859-1')
    concatenado_rs = pd.concat([concatenado_rs, enem_rs], sort = True)

concatenado_rs.to_csv(path+'enem_RS.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

for a in anos:
    enem_mt = pd.read_csv(path+'enem_'+a+'_MT.csv', sep = ';', encoding = 'ISO-8859-1')
    concatenado_mt = pd.concat([concatenado_mt, enem_mt], sort = True)
    
concatenado_mt.to_csv(path+'enem_MT.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

for a in anos:
    enem_df = pd.read_csv(path+'enem_'+a+'_DF.csv', sep = ';', encoding = 'ISO-8859-1')
    concatenado_df = pd.concat([concatenado_df, enem_df], sort = True)
    
concatenado_df.to_csv(path+'enem_DF.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

for a in anos:
    medias = pd.read_csv(path+'medias_'+a+'.csv', sep = ';', encoding = 'ISO-8859-1')
    medias_gerais = pd.concat([medias, medias_gerais], sort = True)
    
medias_gerais.to_csv(path+'medias_gerais.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

for a in anos:
    medias_tpescola = pd.read_csv(path+'medias_tpescola_'+a+'.csv', sep = ';', encoding = 'ISO-8859-1')
    medias_gerais_tpescola = pd.concat([medias_tpescola, medias_gerais_tpescola], sort = True)
    medias_gerais_tpescola.replace(1, 'Federal', inplace = True)
    medias_gerais_tpescola.replace(2, 'Estadual', inplace = True)
    medias_gerais_tpescola.replace(3, 'Municipal', inplace = True)
    medias_gerais_tpescola.replace(4, 'Privada', inplace = True)
    
medias_gerais_tpescola.to_csv(path+'medias_gerais_tpescola.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

ufs = ['RS', 'DF', 'MT']

for u in ufs:
    enem = pd.read_csv(path+'enem_'+ u +'.csv', sep = ';', encoding = 'ISO-8859-1')
    concatenado = pd.concat([concatenado, enem], sort = True)

concatenado.to_csv(path+'enem.csv', sep = ';', index = None, encoding = 'ISO-8859-1')

concatenado.count()

concatenado_df.count()

concatenado_mt.count()

concatenado_rs.count()
