==> 07_identificar rede x concorrentes.ipynb <==
import pandas as pd

path_enem = 'D:/000000001_ENEM/DADOS_PROCESSADOS/enem.csv'
path_unidades = 'D:/000000001_ENEM/DADOS_REDE/UNIDADES_REDE.xlsx'
path_concorrentes = 'D:/000000001_ENEM/DADOS_REDE/CONCORRENTES.xlsx'
path_out = 'D:/000000001_ENEM/DADOS_PROCESSADOS/'

enem = pd.read_csv(path_enem, sep = ';', encoding = 'ISO-8859-1')

enem.head()

unidades_rede = pd.read_excel(path_unidades, encoding = 'ISO-8859-1')
unidades_concorrentes = pd.read_excel(path_concorrentes)

unidades_rede

unidades_rede['ind_rede'] = 1

enem_unidades = pd.merge(enem, unidades_rede, left_on = 'CO_ESCOLA', right_on = 'co_entidade', how = 'left')

enem_unidades.head()

enem_unidades['ind_rede'].fillna(0, inplace = True)

enem_unidades.head()

enem_unidades.groupby(['ind_rede']).count()

unidades_concorrentes ['ind_concorrente'] = 1

unidades_concorrentes

enem_completo = pd.merge(enem_unidades, unidades_concorrentes, left_on = 'CO_ESCOLA', right_on = 'co_entidade', how = 'left')

enem_completo['co_entidade_y'].fillna(0)

enem_completo.head()

enem_completo[enem_completo['ind_concorrente']==1]

enem_completo.rename(columns={'unidade_rede':'nome_entidade_rede','co_entidade_x':'cod_entidade_rede', 'no_entidade':'nome_concorrente','co_entidade_y':'cod_concorrente','unidade_concorrente': 'cod_entidade_concorrente'}, inplace=True)

enem_completo['nome_concorrente'].fillna(0, inplace = True)
enem_completo['cod_concorrente'].fillna(0, inplace = True)
enem_completo['cod_entidade_concorrente'].fillna(0, inplace = True)
enem_completo['nome_entidade_rede'].fillna(0, inplace = True)
enem_completo['cod_entidade_rede'].fillna(0, inplace = True)

enem_completo['chave'] = (enem_completo['NU_ANO'].astype(int)).astype(str) +  '|' +  (enem_completo['CO_ESCOLA'].astype(int)).astype(str)

enem_completo['chave']

enem_completo.to_csv(path_out+'enem_completo.csv', sep = ';', index = None, encoding = 'ISO-8859-1')
