==> 08_escolas_Censo.ipynb <==
import pandas as pd

co_ufs = [43, 53, 51]
## 43 - RS
## 51 - MT
## 53 - DF
path_escolas = 'D:/000000001_ENEM/ESCOLAS/'

escolas_2019 = pd.read_csv(path_escolas+'ESCOLAS_2019.csv', encoding = 'ISO-8859-1', sep = '|', low_memory=False)
for uf in co_ufs:
    path = path_escolas+'escolas_2019_{}.csv'.format(uf)
    escolas_2019[escolas_2019['CO_UF'] == uf].to_csv(path, sep = ';', index = None, encoding = 'ISO-8859-1')

escolas_2018 = pd.read_csv(path_escolas+'ESCOLAS_2018.csv', encoding = 'ISO-8859-1', sep = '|', low_memory=False)
for uf in co_ufs:
    path = path_escolas+'escolas_2018_{}.csv'.format(uf)
    escolas_2018[escolas_2018['CO_UF'] == uf].to_csv(path, sep = ';', index = None, encoding = 'ISO-8859-1')

escolas_2017 = pd.read_csv(path_escolas+'ESCOLAS_2017.csv', encoding = 'ISO-8859-1', sep = '|', low_memory=False)
for uf in co_ufs:
    path = path_escolas+'escolas_2017_{}.csv'.format(uf)
    escolas_2017[escolas_2017['CO_UF'] == uf].to_csv(path, sep = ';', index = None, encoding = 'ISO-8859-1')

escolas_2016 = pd.read_csv(path_escolas+'ESCOLAS_2016.csv', encoding = 'ISO-8859-1', sep = '|', low_memory=False)
for uf in co_ufs:
    path = path_escolas+'escolas_2016_{}.csv'.format(uf)
    escolas_2016[escolas_2016['CO_UF'] == uf].to_csv(path, sep = ';', index = None, encoding = 'ISO-8859-1')

escolas_2015 = pd.read_csv(path_escolas+'ESCOLAS_2015.csv', encoding = 'ISO-8859-1', sep = '|', low_memory=False)
for uf in co_ufs:
    path = path_escolas+'escolas_2015_{}.csv'.format(uf)
    escolas_2015[escolas_2015['CO_UF'] == uf].to_csv(path, sep = ';', index = None, encoding = 'ISO-8859-1')

geral_escolas = pd.concat([escolas_2019, escolas_2018, escolas_2017, escolas_2016, escolas_2015], sort = False)

geral_escolas['chave'] = geral_escolas['NU_ANO_CENSO'].astype(str) +'|'+  geral_escolas['CO_ENTIDADE'].astype(str)

geral_escolas.head()

geral_escolas[['NU_ANO_CENSO','CO_ENTIDADE','NO_ENTIDADE','CO_MUNICIPIO', 'IN_DESKTOP_ALUNO', 'QT_DESKTOP_ALUNO', 'IN_COMP_PORTATIL_ALUNO', 'QT_COMP_PORTATIL_ALUNO', 'IN_TABLET_ALUNO', 'QT_TABLET_ALUNO', 'IN_INTERNET', 'IN_INTERNET_ALUNOS', 'IN_INTERNET_ADMINISTRATIVO', 'IN_INTERNET_APRENDIZAGEM', 'IN_INTERNET_COMUNIDADE', 'IN_ACESSO_INTERNET_COMPUTADOR', 'IN_ACES_INTERNET_DISP_PESSOAIS', 'TP_REDE_LOCAL', 'IN_BANDA_LARGA', 'QT_PROF_PSICOLOGO','QT_PROF_PEDAGOGIA', 'QT_PROF_COORDENADOR', 'IN_MATERIAL_PED_MULTIMIDIA', 'IN_RESERVA_RENDA', 'IN_RESERVA_PPI', 'IN_RESERVA_PUBLICA', 'IN_EXAME_SELECAO' ,'IN_BIBLIOTECA', 'IN_BIBLIOTECA_SALA_LEITURA','IN_LABORATORIO_CIENCIAS', 'IN_LABORATORIO_INFORMATICA', 'IN_EQUIP_LOUSA_DIGITAL', 'QT_EQUIP_LOUSA_DIGITAL', 'IN_EQUIP_MULTIMIDIA', 'QT_EQUIP_MULTIMIDIA', 'chave']].to_csv(path_escolas+'geral_escolas.csv', sep = ';', encoding = 'ISO-8859-1')
