==> 09_Tratamento dados ausentes.ipynb <==
import pandas as pd
import numpy as np

path_enem = 'D:/000000001_ENEM/DADOS_PROCESSADOS/'
path_escolas = 'D:/000000001_ENEM/ESCOLAS/'

enem = pd.read_csv(path_enem+'enem_completo.csv', sep = ';', encoding = 'ISO-8859-1',low_memory=False)

enem.count()

escolas = pd.read_csv(path_escolas+'geral_escolas.csv', sep = ';', encoding = 'ISO-8859-1',low_memory=False)

enem_escolas = pd.merge (enem, escolas, left_on = 'chave', right_on = 'chave', how = 'left')

enem_escolas.to_csv(path_enem + 'base_enem_completa_sem trat.csv', sep = ';', encoding = 'ISO-8859-1')

enem_escolas[['CO_ESCOLA', 'CO_ENTIDADE']].count()

vazios = pd.DataFrame({'Tipos':enem_escolas.dtypes, 'NaN':enem_escolas.isna().sum(),'Total':enem_escolas.count()}).sort_values('NaN', ascending = False)

enem_escolas.columns.values

enem_escolas['TP_ESTADO_CIVIL'].unique()

a = 'TP_ESTADO_CIVIL'
enem_escolas[a].fillna('Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 0, value = 'Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Solteiro', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Casado/Tem companheiro(a)', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Divorciado(a)/Desquitado(a)/Separado(a)', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Viúvo(a)', inplace = True)

enem_escolas['TP_ESTADO_CIVIL'].unique()

enem_escolas['TP_COR_RACA'].unique()

a = 'TP_COR_RACA'
enem_escolas[a].fillna('ND', inplace = True)
enem_escolas[a].replace(to_replace = 0, value = 'ND', inplace = True) ## Não declarado
enem_escolas[a].replace(to_replace = 1, value = 'Branca', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Preta', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Parda', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Amarela', inplace = True)
enem_escolas[a].replace(to_replace = 5, value = 'Indígena', inplace = True)
enem_escolas[a].replace(to_replace = 6, value = 'ND', inplace = True) ## Não dispõe da informação

enem_escolas['TP_COR_RACA'].unique()

enem_escolas['TP_NACIONALIDADE'].unique()

a = 'TP_NACIONALIDADE'
enem_escolas[a].fillna('Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 0, value = 'Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Brasileiro(a)', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Brasileiro(a) Naturalizado(a)', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Estrangeiro(a)', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Brasileiro(a) Nato(a), nascido(a) no exterior', inplace = True)


enem_escolas['TP_NACIONALIDADE'].unique()

enem_escolas['TP_ST_CONCLUSAO'].unique()

a = 'TP_ST_CONCLUSAO'
enem_escolas[a].fillna('Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Já concluí o Ensino Médio', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Estou cursando e concluirei o Ensino Médio no ano corrente', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Estou cursando e concluirei o Ensino Médio após o ano corrente', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Não concluí e não estou cursando o Ensino Médio', inplace = True)

enem_escolas['TP_ST_CONCLUSAO'].unique()

enem_escolas['TP_ESCOLA'].unique()

a = 'TP_ESCOLA'
enem_escolas[a].fillna('Não Respondeu', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Não Respondeu', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Pública', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Privada', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Exterior', inplace = True)

enem_escolas['TP_ESCOLA'].unique()

enem_escolas['TP_ENSINO'].unique()

a = 'TP_ENSINO'
enem_escolas[a].fillna('Não Respondeu', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Ensino Regular', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Educação Especial - Modalidade Substitutiva', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Educação de Jovens e Adultos', inplace = True)

enem_escolas['TP_ENSINO'].unique()

enem_escolas['IN_TREINEIRO'].unique()

a = 'IN_TREINEIRO'
enem_escolas[a].fillna('Não Respondeu', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Sim', inplace = True)
enem_escolas[a].replace(to_replace = 0, value = 'Não', inplace = True)

enem_escolas['IN_TREINEIRO'].unique()

enem_escolas['TP_DEPENDENCIA_ADM_ESC'].unique()

a = 'TP_DEPENDENCIA_ADM_ESC'
enem_escolas[a].fillna('Não Respondeu', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Federal', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Estadual', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Municipal', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Privada', inplace = True)


enem_escolas['TP_DEPENDENCIA_ADM_ESC'].unique()

enem_escolas['TP_LOCALIZACAO_ESC'].unique()

a = 'TP_LOCALIZACAO_ESC'
enem_escolas[a].fillna('Não Respondeu', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Urbana', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Rural', inplace = True)

enem_escolas['TP_LOCALIZACAO_ESC'].unique()

enem_escolas['TP_SIT_FUNC_ESC'].unique()

a = 'TP_SIT_FUNC_ESC'
enem_escolas[a].fillna('Não Informado', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Em atividade', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Paralisada', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Extinta', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Escola extinta em anos anteriores', inplace = True)

enem_escolas['TP_SIT_FUNC_ESC'].unique()

enem_escolas['TP_PRESENCA_CN'].unique()

enem_escolas['TP_PRESENCA_CH'].unique()

enem_escolas['TP_PRESENCA_LC'].unique()

enem_escolas['TP_PRESENCA_MT'].unique()

colunas = ['TP_PRESENCA_CN', 'TP_PRESENCA_CH', 'TP_PRESENCA_LC','TP_PRESENCA_MT']
for c in colunas:
    enem_escolas[c].fillna('Não informado', inplace = True)
    enem_escolas[c].replace(to_replace = 0, value = 'Faltou à prova', inplace = True)
    enem_escolas[c].replace(to_replace = 1, value = 'Presente na prova', inplace = True)
    enem_escolas[c].replace(to_replace = 2, value = 'Eliminado na prova', inplace = True)

enem_escolas['TP_PRESENCA_CN'].unique()

enem_escolas['TP_LINGUA'].unique()

a = 'TP_LINGUA'
enem_escolas[a].fillna('Não Informado', inplace = True)
enem_escolas[a].replace(to_replace = 0, value = 'Inglês', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Espanhol', inplace = True)

enem_escolas['TP_LINGUA'].unique()

enem_escolas['TP_STATUS_REDACAO'].unique()

a = 'TP_STATUS_REDACAO'
enem_escolas[a].fillna('Não declarado', inplace = True)
enem_escolas[a].replace(to_replace = 1, value = 'Sem problemas', inplace = True)
enem_escolas[a].replace(to_replace = 2, value = 'Anulada', inplace = True)
enem_escolas[a].replace(to_replace = 3, value = 'Cópia texto motivador', inplace = True)
enem_escolas[a].replace(to_replace = 4, value = 'Em branco', inplace = True)
enem_escolas[a].replace(to_replace = 5, value = 'Fere Direitos Humanos', inplace = True)
enem_escolas[a].replace(to_replace = 6, value = 'Fuga ao tema', inplace = True)
enem_escolas[a].replace(to_replace = 7, value = 'Não atendimento ao tipo', inplace = True)
enem_escolas[a].replace(to_replace = 8, value = 'Texto insuficiente', inplace = True)
enem_escolas[a].replace(to_replace = 9, value = 'Parte desconectada', inplace = True)
enem_escolas[a].replace(to_replace = 98, value = 'Não atendimento ao item 2.2.5 do edital do exame.', inplace = True)

enem_escolas['TP_STATUS_REDACAO'].unique()

## Q001 - Até que série seu pai, ou o homem responsável por você, estudou?
## Q002 - Até que série sua mãe, ou a mulher responsável por você, estudou?
## Q003 - A partir da apresentação de algumas ocupações divididas em grupos ordenados, indique o grupo que contempla a ocupação mais próxima da ocupação do seu pai ou do homem responsável por você. (Se ele não estiver trabalhando, escolha uma ocupação pensando no último trabalho dele).
## Q004 - A partir da apresentação de algumas ocupações divididas em grupos ordenados, indique o grupo que contempla a ocupação mais próxima da ocupação da sua mãe ou da mulher responsável por você. (Se ela não estiver trabalhando, escolha uma ocupação pensando no último trabalho dela).
## Q005 - Incluindo você, quantas pessoas moram atualmente em sua residência?
## Q006 - Qual é a renda mensal de sua família? (Some a sua renda com a dos seus familiares.)
## Q007 - Em sua residência trabalha empregado(a) doméstico(a)?
## Q008 - Na sua residência tem banheiro?
## Q009 - Na sua residência tem quartos para dormir?
## Q010 - Na sua residência tem carro?
## Q011 - Na sua residência tem motocicleta?
## Q012 - Na sua residência tem geladeira?
## Q013 - Na sua residência tem freezer (independente ou segunda porta da geladeira)?
## Q014 - Na sua residência tem máquina de lavar roupa? (o tanquinho NÃO deve ser considerado)
## Q015 - Na sua residência tem máquina de secar roupa (independente ou em conjunto com a máquina de lavar roupa)?
## Q016 - Na sua residência tem forno micro-ondas?
## Q017 - Na sua residência tem máquina de lavar louça?
## Q018 - Na sua residência tem aspirador de pó?
## Q019 - Na sua residência tem televisão em cores?
## Q020 - Na sua residência tem aparelho de DVD?
## Q021 - Na sua residência tem TV por assinatura?
## Q022 - Na sua residência tem telefone celular?
## Q023 - Na sua residência tem telefone fixo?
## Q024 - Na sua residência tem computador?
## Q025 - Na sua residência tem acesso à Internet?


enem_escolas['Q001'].unique()

enem_escolas['Q002'].unique()

q = ['Q001','Q002']
for a in q:
    enem_escolas[a].fillna('Não informado', inplace = True)
    enem_escolas[a].replace(to_replace = 'A', value = 'Nunca estudou', inplace = True)
    enem_escolas[a].replace(to_replace = 'B', value = 'Não completou a 4ª série/5º ano do Ensino Fundamental', inplace = True)
    enem_escolas[a].replace(to_replace = 'C', value = 'Completou a 4ª série/5º ano, mas não completou a 8ª série/9º ano do Ensino Fundamental', inplace = True)
    enem_escolas[a].replace(to_replace = 'D', value = 'Completou a 8ª série/9º ano do Ensino Fundamental, mas não completou o Ensino Médio', inplace = True)
    enem_escolas[a].replace(to_replace = 'E', value = 'Completou o Ensino Médio, mas não completou a Faculdade', inplace = True)
    enem_escolas[a].replace(to_replace = 'F', value = 'Completou a Faculdade, mas não completou a Pós-graduação', inplace = True)
    enem_escolas[a].replace(to_replace = 'G', value = 'Completou a Pós-graduação', inplace = True)
    enem_escolas[a].replace(to_replace = 'H', value = 'Não sei.', inplace = True)

enem_escolas['Q001'].unique()

enem_escolas['Q002'].unique()

##['Q003','Q004']
## Grupo 1: Lavradora, agricultora sem empregados, bóia fria, criadora de animais (gado, porcos, galinhas, ovelhas, cavalos etc.), apicultora, pescadora, lenhadora, seringueira, extrativista.
## Grupo 2: Diarista, empregada doméstica, cuidadora de idosos, babá, cozinheira (em casas particulares), motorista particular, jardineira, faxineira de empresas e prédios, vigilante, porteira, carteira, office-boy, vendedora, caixa, atendente de loja, auxiliar administrativa, recepcionista, servente de pedreiro, repositora de mercadoria.
## Grupo 3: Padeira, cozinheira industrial ou em restaurantes, sapateira, costureira, joalheira, torneira mecânica, operadora de máquinas, soldadora, operária de fábrica, trabalhadora da mineração, pedreira, pintora, eletricista, encanadora, motorista, caminhoneira, taxista.
## Grupo 4: Professora (de ensino fundamental ou médio, idioma, música, artes etc.), técnica (de enfermagem, contabilidade, eletrônica etc.), policial, militar de baixa patente (soldado, cabo, sargento), corretora de imóveis, supervisora, gerente, mestre de obras, pastora, microempresária (proprietária de empresa com menos de 10 empregados), pequena comerciante, pequena proprietária de terras, trabalhadora autônoma ou por conta própria.
## Grupo 5: Médica, engenheira, dentista, psicóloga, economista, advogada, juíza, promotora, defensora, delegada, tenente, capitã, coronel, professora universitária, diretora em empresas públicas ou privadas, política, proprietária de empresas com mais de 10 empregados.
## Não sei.


enem_escolas['Q003'].unique()

enem_escolas['Q004'].unique()

q = ['Q003','Q004']
for a in q:
    enem_escolas[a].fillna('Não informado', inplace = True)
    enem_escolas[a].replace(to_replace = 'A', value = 'Grupo 1', inplace = True)
    enem_escolas[a].replace(to_replace = 'B', value = 'Grupo 2', inplace = True)
    enem_escolas[a].replace(to_replace = 'C', value = 'Grupo 3', inplace = True)
    enem_escolas[a].replace(to_replace = 'D', value = 'Grupo 4', inplace = True)
    enem_escolas[a].replace(to_replace = 'E', value = 'Grupo 5', inplace = True)
    enem_escolas[a].replace(to_replace = 'F', value = 'Não sei.', inplace = True)

enem_escolas['Q003'].unique()

enem_escolas['Q004'].unique()

enem_escolas['Q005'].unique()

enem_escolas['Q005'].fillna('Não informado', inplace = True)

enem_escolas['Q006'].unique()

a = 'Q006'
enem_escolas[a].fillna('Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 'A', value = 'Nenhuma renda', inplace = True)
enem_escolas[a].replace(to_replace = 'B', value = 'Até R$ 998,00', inplace = True)
enem_escolas[a].replace(to_replace = 'C', value = 'De R$ 998,01 até R$ 1.497,00', inplace = True)
enem_escolas[a].replace(to_replace = 'D', value = 'De R$ 1.497,01 até R$ 1.996,00', inplace = True)
enem_escolas[a].replace(to_replace = 'E', value = 'De R$ 1.996,01 até R$ 2.495,00', inplace = True)
enem_escolas[a].replace(to_replace = 'F', value = 'De R$ 2.495,01 até R$ 2.994,00', inplace = True)
enem_escolas[a].replace(to_replace = 'G', value = 'De R$ 2.994,01 até R$ 3.992,00', inplace = True)
enem_escolas[a].replace(to_replace = 'H', value = 'De R$ 3.992,01 até R$ 4.990,00', inplace = True)
enem_escolas[a].replace(to_replace = 'I', value = 'De R$ 4.990,01 até R$ 5.988,00', inplace = True)
enem_escolas[a].replace(to_replace = 'J', value = 'De R$ 5.988,01 até R$ 6.986,00', inplace = True)
enem_escolas[a].replace(to_replace = 'K', value = 'De R$ 6.986,01 até R$ 7.984,00', inplace = True)
enem_escolas[a].replace(to_replace = 'L', value = 'De R$ 7.984,01 até R$ 8.982,00', inplace = True)
enem_escolas[a].replace(to_replace = 'M', value = 'De R$ 8.982,01 até R$ 9.980,00', inplace = True)
enem_escolas[a].replace(to_replace = 'N', value = 'De R$ 9.980,01 até R$ 11.976,00', inplace = True)
enem_escolas[a].replace(to_replace = 'O', value = 'De R$ 11.976,01 até R$ 14.970,00', inplace = True)
enem_escolas[a].replace(to_replace = 'P', value = 'De R$ 14.970,01 até R$ 19.960,00', inplace = True)
enem_escolas[a].replace(to_replace = 'Q', value = 'Mais de R$ 19.960,00.', inplace = True)


enem_escolas['Q006'].unique()

enem_escolas['Q007'].unique()

a = 'Q007'
enem_escolas[a].fillna('Não informado', inplace = True)
enem_escolas[a].replace(to_replace = 'A', value = 'Não', inplace = True)
enem_escolas[a].replace(to_replace = 'B', value = 'Sim, um ou dois dias por semana', inplace = True)
enem_escolas[a].replace(to_replace = 'C', value = 'Sim, três ou quatro dias por semana', inplace = True)
enem_escolas[a].replace(to_replace = 'D', value = 'Sim, pelo menos cinco dias por semana', inplace = True)

enem_escolas['Q007'].unique()

q = ['Q008', 'Q009', 'Q010', 'Q011', 'Q012', 'Q013', 'Q014','Q015', 'Q016', 'Q017','Q019', 'Q022', 'Q024']
for a in q:
    print(a)
    print(enem_escolas[a].unique())

q = ['Q008', 'Q009', 'Q010', 'Q011', 'Q012', 'Q013', 'Q014','Q015', 'Q016', 'Q017','Q019', 'Q022', 'Q024']
for a in q:
    enem_escolas[a].fillna('Não informado', inplace = True)
    enem_escolas[a].replace(to_replace = 'A', value = 'Não', inplace = True)
    enem_escolas[a].replace(to_replace = 'B', value = 'Sim, um', inplace = True)
    enem_escolas[a].replace(to_replace = 'C', value = 'Sim, dois', inplace = True)
    enem_escolas[a].replace(to_replace = 'D', value = 'Sim, três', inplace = True)
    enem_escolas[a].replace(to_replace = 'E', value = 'Sim, quatro ou mais', inplace = True)

q = ['Q008', 'Q009', 'Q010', 'Q011', 'Q012', 'Q013', 'Q014','Q015', 'Q016', 'Q017','Q019', 'Q022', 'Q024']
for a in q:
    print(a)
    print(enem_escolas[a].unique())

q = [ 'Q018', 'Q020', 'Q021','Q023','Q025']
for a in q:
    print(a)
    print(enem_escolas[a].unique())

q = [ 'Q018', 'Q020', 'Q021','Q023','Q025']
for a in q:
    enem_escolas[a].fillna('Não informado', inplace = True)
    enem_escolas[a].replace(to_replace = 'A', value = 'Não', inplace = True)
    enem_escolas[a].replace(to_replace = 'B', value = 'Sim', inplace = True)   

q = [ 'Q018', 'Q020', 'Q021','Q023','Q025']
for a in q:
    print(a)
    print(enem_escolas[a].unique())

q = [  'ind_rede', 'ind_concorrente','IN_DESKTOP_ALUNO', 'IN_COMP_PORTATIL_ALUNO','IN_TABLET_ALUNO', 'IN_INTERNET',
     'IN_INTERNET_ALUNOS', 'IN_INTERNET_ADMINISTRATIVO','IN_INTERNET_APRENDIZAGEM', 'IN_INTERNET_COMUNIDADE',
     'IN_ACESSO_INTERNET_COMPUTADOR', 'IN_ACES_INTERNET_DISP_PESSOAIS','TP_REDE_LOCAL', 'IN_BANDA_LARGA',
     'IN_MATERIAL_PED_MULTIMIDIA', 'IN_RESERVA_RENDA', 'IN_RESERVA_PPI','IN_RESERVA_PUBLICA', 'IN_EXAME_SELECAO',
     'IN_BIBLIOTECA', 'IN_BIBLIOTECA_SALA_LEITURA', 'IN_LABORATORIO_CIENCIAS', 'IN_LABORATORIO_INFORMATICA','IN_EQUIP_LOUSA_DIGITAL',
     'IN_EQUIP_MULTIMIDIA']
for a in q:
    print(a)
    print(enem_escolas[a].unique())

q = [ 'ind_rede', 'ind_concorrente','IN_DESKTOP_ALUNO', 'IN_COMP_PORTATIL_ALUNO','IN_TABLET_ALUNO', 'IN_INTERNET',
     'IN_INTERNET_ALUNOS', 'IN_INTERNET_ADMINISTRATIVO','IN_INTERNET_APRENDIZAGEM', 'IN_INTERNET_COMUNIDADE',
     'IN_ACESSO_INTERNET_COMPUTADOR', 'IN_ACES_INTERNET_DISP_PESSOAIS', 'IN_BANDA_LARGA',
     'IN_MATERIAL_PED_MULTIMIDIA', 'IN_RESERVA_RENDA', 'IN_RESERVA_PPI','IN_RESERVA_PUBLICA', 'IN_EXAME_SELECAO',
     'IN_BIBLIOTECA', 'IN_BIBLIOTECA_SALA_LEITURA', 'IN_LABORATORIO_CIENCIAS', 'IN_LABORATORIO_INFORMATICA','IN_EQUIP_LOUSA_DIGITAL',
     'IN_EQUIP_MULTIMIDIA']
for a in q:
    enem_escolas[a].fillna('Não', inplace = True)
    enem_escolas[a].replace(to_replace = 0, value = 'Não', inplace = True)
    enem_escolas[a].replace(to_replace = 1, value = 'Sim', inplace = True)
    enem_escolas[a].replace(to_replace = 9, value = 'Não informado', inplace = True)

q = [  'ind_rede', 'ind_concorrente','IN_DESKTOP_ALUNO', 'IN_COMP_PORTATIL_ALUNO','IN_TABLET_ALUNO', 'IN_INTERNET',
     'IN_INTERNET_ALUNOS', 'IN_INTERNET_ADMINISTRATIVO','IN_INTERNET_APRENDIZAGEM', 'IN_INTERNET_COMUNIDADE',
     'IN_ACESSO_INTERNET_COMPUTADOR', 'IN_ACES_INTERNET_DISP_PESSOAIS', 'IN_BANDA_LARGA',
     'IN_MATERIAL_PED_MULTIMIDIA', 'IN_RESERVA_RENDA', 'IN_RESERVA_PPI','IN_RESERVA_PUBLICA', 'IN_EXAME_SELECAO',
     'IN_BIBLIOTECA', 'IN_BIBLIOTECA_SALA_LEITURA', 'IN_LABORATORIO_CIENCIAS', 'IN_LABORATORIO_INFORMATICA','IN_EQUIP_LOUSA_DIGITAL',
     'IN_EQUIP_MULTIMIDIA']
for a in q:
    print(a)
    print(enem_escolas[a].unique())

enem_escolas['TP_REDE_LOCAL'].unique()

a = 'TP_REDE_LOCAL'
enem_escolas[a].fillna('Não', inplace = True)
enem_escolas[a].replace(to_replace = 0, value = 'Não', inplace = True) ## Não há rede local interligando computadores
enem_escolas[a].replace(to_replace = 1, value = 'Sim', inplace = True) ## A cabo
enem_escolas[a].replace(to_replace = 2, value = 'Sim', inplace = True) ## Wireless
enem_escolas[a].replace(to_replace = 3, value = 'Sim', inplace = True) ## A cabo e Wireless
enem_escolas[a].replace(to_replace = 9, value = 'Não informado', inplace = True)

enem_escolas['TP_REDE_LOCAL'].unique()

q = ['QT_PROF_PEDAGOGIA','QT_TABLET_ALUNO', 'QT_PROF_COORDENADOR', 'QT_EQUIP_LOUSA_DIGITAL', 'QT_DESKTOP_ALUNO',
    'QT_COMP_PORTATIL_ALUNO','QT_PROF_PSICOLOGO', 'QT_EQUIP_MULTIMIDIA' ]
for a in q:
    enem_escolas[a].fillna(0, inplace = True)

vazios = pd.DataFrame({'Tipos':enem_escolas.dtypes, 'NaN':enem_escolas.isna().sum(),'Total':enem_escolas.count()}).sort_values('NaN', ascending = False)

vazios[vazios['NaN']>0]

enem_escolas.dropna(axis = 0, how = 'all', inplace=True, subset= ['CO_ENTIDADE'])
enem_escolas.dropna(axis = 0, how = 'all', inplace=True, subset= [ 'NU_IDADE'])

vazios = pd.DataFrame({'Tipos':enem_escolas.dtypes, 'NaN':enem_escolas.isna().sum(),'Total':enem_escolas.count()}).sort_values('NaN', ascending = False)

vazios[vazios['NaN']>0]

enem_escolas.to_csv(path_enem+'base_enem_completa.csv', sep = ';', encoding = 'ISO-8859-1', index = None)
